+++ README.md
@@ -1,3 +1,13 @@
+# Cross-Platform Media Player
+
+This can play HLS, DASH and SmoothStreaming files across all platforms.
+
+For **iOS**, there are limitations:
+- The URL should end with `.mpd`
+- Its format should match as described in **MPD_DASH_FORMAT.md**
+
+---
+
 ## 🙌 Inspiration & Credits
 
 ### Inspired by:
@@ -14,6 +24,9 @@
 - [open-ani/mediamp (mediamp-vlc)](https://github.com/open-ani/mediamp/tree/main/mediamp-vlc)  
   Thanks to this, **Desktop support** is now possible without requiring VLC to be installed on the system — VLC binaries are shipped with the application itself.
 
+- [MPEGDASHAVPlayerDemo](https://github.com/h-anders-unext/MPEGDASHAVPlayerDemo)  
+  Thanks to this, I got some idea to parse DASH files.
+
 ---
 
 ## 🎬 Demo
@@ -24,8 +37,11 @@
 **Desktop Demo:**  
 ![Desktop Demo](others/demo/desktop.gif)
 
+**iOS Demo:**  
+![iOS Demo](others/demo/ios.gif)
+
 **Android Demo:**  
-![Desktop Demo](others/demo/android.gif)
+![Android Demo](others/demo/android.gif)
 
 ---
 
@@ -34,11 +50,11 @@
 - Kotlin/JS
 - Kotlin/Desktop (**bundled VLC binaries — no external installation needed**)
 - Android/Phone
+- iOS
 
 ---
 
 ## 🚧 Pending Platform Targets
-- iOS
 - TV
 
 ---
@@ -49,6 +65,7 @@
 - Pause
 - Dispose
 - Change media
+
 ---
 
 ## 🔜 Pending Features
@@ -59,7 +76,7 @@
 ---
 
 ## ⏭️ Next Upcoming Target
-- **TV/IOS**
+- **Desktop (With own parsing, without depending on VLC, may be VLC can be removed so that i can move back to Apache licence)**
 
 ---
 
@@ -68,34 +85,32 @@
 Desktop support ships with **pre-bundled VLC binaries**, meaning it works even if VLC is not installed on the user’s machine.
 
 ### Example folder structure:
-   ```
-   desktopApp/
-    |- appResources/
-    |  |- macos-arm64/
-    |  |  |- lib/
-    |  |  |  |- libvlc.dylib
-    |  |  |- plugins/
-    |  |  |  |- xxx.dylib
-    |  |- macos-x64/
-    |  |  |- lib/
-    |  |  |  |- libvlc.dylib
-    |  |  |- plugins/
-    |  |  |  |- xxx.dylib
-    |  |- windows-x64/
-    |  |  |- lib/
-    |  |  |  |- libvlc.dll
-    |  |  |  |- plugins/
-    |  |- linux-x64/
-    |  |  |- lib/
-    |  |  |  |- libvlc.so
-    |- src/
-    |- build.gradle.kts
-   ```
+```
+desktopApp/
+ |- appResources/
+ |  |- macos-arm64/
+ |  |  |- lib/
+ |  |  |  |- libvlc.dylib
+ |  |  |- plugins/
+ |  |  |  |- xxx.dylib
+ |  |- macos-x64/
+ |  |  |- lib/
+ |  |  |  |- libvlc.dylib
+ |  |  |- plugins/
+ |  |  |  |- xxx.dylib
+ |  |- windows-x64/
+ |  |  |- lib/
+ |  |  |  |- libvlc.dll
+ |  |  |  |- plugins/
+ |  |- linux-x64/
+ |  |  |- lib/
+ |  |  |  |- libvlc.so
+ |- src/
+ |- build.gradle.kts
+```
 
 ### Notes:
-- It is in very early stage, so mostly are for development purposes, once all platform integrated, will deep dive into proper functional
+- It is in very early stage, so mostly for development purposes. Once all platforms are integrated, we’ll deep dive into proper functionality.
 - Currently, **macOS binaries** are included. You can extend this to other platforms by adding respective VLC builds.
 - On the **first run**, VLC may take ~2 minutes to cache/generate `plugins.dat`.  
   From the **second run onward**, startup time reduces to ~7–10 seconds.
-
----
